Partial revert of 4820.
