+++ cli/internal/ffi/bindings.h
@@ -18,8 +18,6 @@ struct Buffer previous_content(struct Buffer buffer);
 
 struct Buffer recursive_copy(struct Buffer buffer);
 
-struct Buffer get_package_file_hashes_from_git_index(struct Buffer buffer);
-
 struct Buffer transitive_closure(struct Buffer buf);
 
 struct Buffer subgraph(struct Buffer buf);
