+++ crates/turborepo-ffi/messages.proto
@@ -149,19 +149,3 @@ message RecursiveCopyRequest {
 message RecursiveCopyResponse {
   optional string error = 1;
 }
-
-message GetPackageFileHashesFromGitIndexRequest {
-  string turbo_root = 1;
-  string package_path = 2;
-}
-
-message FileHashes {
-  map<string, string> hashes = 1;
-}
-
-message GetPackageFileHashesFromGitIndexResponse {
-  oneof response {
-    FileHashes hashes = 1;
-    string error = 2;
-  }
-}
