--- crates/turborepo-lib/.gitignore
@@ -1,1 +1,0 @@
-file_descriptor_set.bin
