+++ crates/turborepo-scm/Cargo.toml
@@ -7,9 +7,9 @@ license = "MPL-2.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = { workspace = true }
 dunce = { workspace = true }
 git2 = { version = "0.16.1", default-features = false }
-nom = "7.1.3"
 thiserror = { workspace = true }
 turbopath = { workspace = true }
