--- crates/turborepo-scm/fixtures/01-git-hash-object/.gitignore
@@ -1,2 +1,0 @@
-"quote"
-new*line
