+++ crates/turborepo-scm/src/lib.rs
@@ -2,16 +2,12 @@
 #![feature(provide_any)]
 #![feature(assert_matches)]
 
-use std::backtrace::{self, Backtrace};
+use std::backtrace;
 
 use thiserror::Error;
 use turbopath::PathError;
 
 pub mod git;
-mod hash_object;
-mod ls_tree;
-pub mod package_deps;
-mod status;
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -23,15 +19,4 @@ pub enum Error {
     Io(#[from] std::io::Error, #[backtrace] backtrace::Backtrace),
     #[error("path error: {0}")]
     Path(#[from] PathError, #[backtrace] backtrace::Backtrace),
-    #[error("encoding error: {0}")]
-    Encoding(
-        #[from] std::string::FromUtf8Error,
-        #[backtrace] backtrace::Backtrace,
-    ),
-}
-
-impl Error {
-    pub(crate) fn git_error(s: impl Into<String>) -> Self {
-        Error::Git(s.into(), Backtrace::capture())
-    }
 }
